removed 'xxx' stuff and minor cleaning

+++ groups/btl/btlsos/btlsos_tcptimedcbconnector.cpp
@@ -777,7 +777,7 @@ void TcpTimedCbConnector::cancelAll() {
         d_callbacks.erase(d_callbacks.begin(),
                           d_callbacks.begin() + d_callbacks.size() - 1);
         BSLS_ASSERT(d_currentRequest_p == d_callbacks.back());
-        int numToCancel = toBeCancelled.size();
+        int numToCancel = static_cast<int>(toBeCancelled.size());
         while (--numToCancel >= 0) {
             TcpTimedCbConnector_Reg *reg = toBeCancelled[numToCancel];
             reg->invoke(-1);
@@ -791,7 +791,7 @@ void TcpTimedCbConnector::cancelAll() {
         bsl::deque<TcpTimedCbConnector_Reg *> toBeCancelled(
                                                    d_callbacks, d_allocator_p);
         d_callbacks.clear();
-        int numToCancel = toBeCancelled.size();
+        int numToCancel = static_cast<int>(toBeCancelled.size());
         if (numToCancel) {
             BSLS_ASSERT(d_connectingSocket_p);
             d_manager_p->deregisterSocketEvent(
